+++ core/chain_makers.go
@@ -34,7 +34,7 @@ import (
 	"github.com/ava-labs/coreth/consensus"
 	"github.com/ava-labs/coreth/core/extstate"
 	"github.com/ava-labs/coreth/params"
-	"github.com/ava-labs/coreth/plugin/evm/header"
+	"github.com/ava-labs/coreth/plugin/evm/customheader"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/consensus/misc/eip4844"
 	"github.com/ava-labs/libevm/core/rawdb"
@@ -376,11 +376,11 @@ func (cm *chainMaker) makeHeader(parent *types.Block, gap uint64, state *state.S
 	time := parent.Time() + gap // block time is fixed at [gap] seconds
 
 	config := params.GetExtra(cm.config)
-	gasLimit, err := header.GasLimit(config, parent.Header(), time)
+	gasLimit, err := customheader.GasLimit(config, parent.Header(), time)
 	if err != nil {
 		panic(err)
 	}
-	baseFee, err := header.BaseFee(config, parent.Header(), time)
+	baseFee, err := customheader.BaseFee(config, parent.Header(), time)
 	if err != nil {
 		panic(err)
 	}
