+++ plugin/evm/customheader/block_gas_cost.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package header
+package customheader
 
 import (
 	"errors"
