+++ plugin/evm/imports_test.go
@@ -57,12 +57,12 @@ func TestMustNotImport(t *testing.T) {
 		// possible to do so for both coreth and subnet-evm, where the client may
 		// wish to connect to multiple chains.
 
-		"plugin/evm/atomic": {"core", "plugin/evm/customtypes", "core/extstate", "params"},
-		"plugin/evm/client": {"core", "plugin/evm/customtypes", "core/extstate", "params"},
-		"plugin/evm/config": {"core", "plugin/evm/customtypes", "core/extstate", "params"},
-		"plugin/evm/header": {"core", "core/extstate", "core/vm", "params"},
-		"ethclient":         {"plugin/evm/customtypes", "core/extstate", "params"},
-		"warp":              {"plugin/evm/customtypes", "core/extstate", "params"},
+		"plugin/evm/atomic":       {"core", "plugin/evm/customtypes", "core/extstate", "params"},
+		"plugin/evm/client":       {"core", "plugin/evm/customtypes", "core/extstate", "params"},
+		"plugin/evm/config":       {"core", "plugin/evm/customtypes", "core/extstate", "params"},
+		"plugin/evm/customheader": {"core", "core/extstate", "core/vm", "params"},
+		"ethclient":               {"plugin/evm/customtypes", "core/extstate", "params"},
+		"warp":                    {"plugin/evm/customtypes", "core/extstate", "params"},
 	}
 
 	for packageName, forbiddenImports := range mustNotImport {
